Add PimUniqueApplication::activateApplication
Summary:
This allows one PIM application to ensure another one is started, either
standalone or in kontact, and in the latter case, select the
corresponding plugin to make it visible.

This implementation relies on DBus activation, recently added for the
PIM apps.

+ cleanup unused "bool unique = true" in start() method, they are all unique
+ fix docu

Test Plan: see kdepim-addons change request using this

Reviewers: vkrause, dvratil, mlaurent

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D22621